+++ src/Lib/Engine/Admin/AdminEngineInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PP\Lib\Engine\Admin;
+
+interface AdminEngineInterface {
+
+}